test: move 관련 테스트 패키지 분리
